requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
google-cloud-storage>=2.18.0
google-cloud-firestore>=2.19.0
google-cloud-tasks>=2.17.0
google-cloud-aiplatform>=1.75.0
google-auth>=2.36.0
requests>=2.32.0
